refactor: move K3 chat request validation/thinking out of api_server
api_server.py is the ASGI wiring layer, not a home for chat-specific
utilities. Move validate_chat_request / resolve_thinking (and their
private helpers _validate_one_tool / _validate_tool_list and the
tool-name/effort/tool-choice constants) into serving_chat.py, which
already owns the chat request/response helpers and imports both protocol
and reasoning. api_server now imports the two public entry points and
drops its now-unused `re` import. Behavior is unchanged.
